fix: `authOptions` is not a valid export field for route
